requirements.txt
fastapi
uvicorn
transformers
google-genai
torch
pydantic
safetensors
